requirements.txt
jango==2.2.13django-ckeditor==5.9.0django-debug-toolbar==2.2django-js-asset==1.2.2Pillow==7.1.2pytz==2020.1sqlparse==0.3.1black==19.3b0pylint==2.3.1